Support minimal roll logic
This will eventually become a slash command, but for now, we can simply
resond to a 'special' message with an exclamation mark on it.

```
!roll 20
```

Will respond with a number between 1 and 20. The future slash command
will become:

```
/roll d20
```

perhaps with an alias for `d20` to `20`, but also supporting other
random number generation means, such as ranges or even names from a
corpus.