Support subtraction
Not quite a packrat parser, but this is still muddling lexing and
parsing into the same pass. However, for our purposes and time
constraint, this is fine. We could eventually separate things out to
make it a bit clearer, and for whitespace oriented languages there are
cool tricks like putting lexical information on the tokens to better
drive the decision for such a refactor.