Parse parenthesis
This supports parsing parentheticals to later be used in evaluation, but
still builds up implicit parenthesis on the wrong side (right hand side)
of the production. What is more natural is that and expression such as,

```
1 + 2 + 3
```

turns into,

```
(1 + 2) + 3
```

The lingering question is if we should flatten the explicit and implicit
versions, such that the above has the same AST as the bottom if both
sources are presented ot the parser.