Support parenthesis
This supports parenthesis but for now doesn't mean much as addition and
subtraction have equal precedence. If we add factors such as
multiplication and division then we can better utilise this
functionality. In that case, the evaluator must change and always
evaluate the term of precedence first if it differs from the default
(left) factor.

This is a bit neater in general than what was there before, which was to
add an implicit paren term on the right hand side to make things work.
It changes the AST to express binops as Term `op` Expr pairs which more
closely matches the grammar written in documentation.