Add roundtrip law test
This is more experimental than serious for adding. It is iffy generating
recursive data structures with something like this where we can't
sensibly bound the size of the AST we generate. We could if we had a
means to continually reduce the generators size as we traverse the
stack, or provide an additional context that allows this. Technically
something like an Arbitrary for a BoundedExpr could work, in which the
size is reduce to a particular depth at each recursive function
invocation. Then the base cases are either leaves of the tree or, when
we've hit the maximum depth we're allowed to go, we only veer towards
leave nodes and prevent further recursion.